+++ cmd/root.go
@@ -97,6 +97,7 @@ func checkDefaultsFile() {
 	defaults.LoadConfiguration()
 
 	shellPath, _ := flags.GetString(globals.ShellPathLabel)
+	mysqlshPath, _ := flags.GetString(globals.MysqlshPathLabel)
 
 	shellPath, err := common.GetBashPath(shellPath)
 	if err != nil {
@@ -104,6 +105,7 @@ func checkDefaultsFile() {
 	}
 	if defaults.ValidateDefaults(defaults.Defaults()) {
 		defaults.UpdateDefaults(globals.ShellPathLabel, shellPath, false)
+		defaults.UpdateDefaults(globals.MysqlshPathLabel, mysqlshPath, false)
 	}
 	err = sandbox.FillMockTemplates()
 	if err != nil {
@@ -149,6 +151,7 @@ func init() {
 	setPflag(rootCmd, globals.SandboxHomeLabel, "", "SANDBOX_HOME", defaults.Defaults().SandboxHome, "Sandbox deployment directory", false)
 	setPflag(rootCmd, globals.SandboxBinaryLabel, "", "SANDBOX_BINARY", defaults.Defaults().SandboxBinary, "Binary repository", false)
 	setPflag(rootCmd, globals.ShellPathLabel, "", "SHELL_PATH", common.Which("bash"), "Path to Bash, used for generated scripts", false)
+	setPflag(rootCmd, globals.MysqlshPathLabel, "", "MYSQLSH_PATH", "mysqlsh", "Path to mysqlsh executable", false)
 	rootCmd.PersistentFlags().BoolP(globals.SkipLibraryCheck, "", false, "Skip check for needed libraries (may cause nasty errors)")
 
 	rootCmd.InitDefaultVersionFlag()