+++ cmd/single.go
@@ -357,6 +357,7 @@ func fillSandboxDefinition(cmd *cobra.Command, args []string, usingImport bool) 
 	sd.InitGeneralLog, _ = flags.GetBool(globals.InitGeneralLogLabel)
 	sd.EnableGeneralLog, _ = flags.GetBool(globals.EnableGeneralLogLabel)
 	sd.ShellPath = defaults.Defaults().ShellPath
+	sd.MysqlshPath = defaults.Defaults().MysqlshPath
 
 	if sd.DisableMysqlX && sd.EnableMysqlX {
 		common.Exit(1, "flags --enable-mysqlx and --disable-mysqlx cannot be used together")